+++ RSDKv3/Ini.cpp
@@ -155,7 +155,7 @@ int IniParser::SetString(const char *section, const char *key, char *value)
         }
     }
     if (where < 0) {
-        where = items.size();
+        where = (int)items.size();
         items.push_back(ConfigItem());
     }
 
@@ -177,7 +177,7 @@ int IniParser::SetInteger(const char *section, const char *key, int value)
         }
     }
     if (where < 0) {
-        where = items.size();
+        where = (int)items.size();
         items.push_back(ConfigItem());
     }
 
@@ -199,7 +199,7 @@ int IniParser::SetFloat(const char *section, const char *key, float value)
         }
     }
     if (where < 0) {
-        where = items.size();
+        where = (int)items.size();
         items.push_back(ConfigItem());
     }
 
@@ -221,7 +221,7 @@ int IniParser::SetBool(const char *section, const char *key, bool value)
         }
     }
     if (where < 0) {
-        where = items.size();
+        where = (int)items.size();
         items.push_back(ConfigItem());
     }
 
@@ -243,7 +243,7 @@ int IniParser::SetComment(const char *section, const char *key, const char *comm
         }
     }
     if (where < 0) {
-        where = items.size();
+        where = (int)items.size();
         items.push_back(ConfigItem());
     }
